Add typings for create account operation.
